+++ crun.sh
@@ -11,16 +11,58 @@ CRUN_URL="https://github.com/GochoMugo/crun"
 function show_help() {
     echo " usage:"
     echo "    as a shebang in your C file: #!/usr/bin/env crun"
-    echo "    direct invocation:           crun filename.c"
-    echo "                                 crun <options>"
+    echo "    direct invocation:"
+    echo "          crun filename.c"
+    echo "          crun --create|create-force <path>"
+    echo "          crun --help|version"
     echo
     echo " options:"
-    echo "    -h, --help        show this help information and exit"
-    echo "    -v, --version     show version information"
+    echo "    -c,  --create        create a template script"
+    echo "    -cf, --create-force  same as '-c', but allow overwriting"
+    echo "    -h,  --help          show this help information and exit"
+    echo "    -v,  --version       show version information"
     echo
     echo " see ${CRUN_URL} for feature requests and bug reporting"
     echo
 }
+
+
+# simple templating for a new C file
+# ${1} - file path
+# ${2} - whether to force creation i.e. clobber
+function new() {
+    local force="${2:-}"
+
+    if [ ! "${1}" ]
+    then
+        echo "ERROR: NO path provided"
+        return 1
+    fi
+
+    if [ -f "${1}" ]
+    then
+        echo "WARNING: file found at '${1}'"
+        if [ ! "${force}" ]
+        then
+            echo "ERROR: stopping to avoid overwriting existing file"
+            echo "ERROR: use '--create-force' to force clobbering"
+            return 1
+        else
+            echo "WARNING: file overwritten"
+        fi
+    fi
+
+    {
+        echo "#!/usr/bin/env crun"
+        echo "/* -Wall -O3 */"
+        echo
+        echo "int main(void) {"
+        echo "  return 0;"
+        echo "}"
+    } > "${1}"
+    chmod u+x "${1}"
+}
+
 
 # some arguments processing
 case ${1:-''} in
@@ -32,11 +74,19 @@ case ${1:-''} in
         show_help
         exit
     ;;
+    "-c" | "--create" )
+        new "${2}"
+        exit $?
+    ;;
+    "-cf" | "--create-force" )
+        new "${2}" "force"
+        exit $?
+    ;;
 esac
 
 # ensure we have a path to an existing source file
 [ -f "${1}" ] || {
-    echo " ERROR: NO file path provided"
+    echo " ERROR: NO regular file at the path provided"
     echo
     show_help
     exit 1